Update README.md
The smart store project enables a hassle free and faster shopping, caters to the safety of the customers and even contributes in reducing the carbon foot print. It aims at providing a more comfortable and enjoyable shopping experience. This is mainly beneficial to the customers who procure less commodities and are in a hurry to leave.
The features implemented are - smart trolley, automated lights, door and fire alarm and an system to count and display the number of customers present in a billing line.